Fix FindSqlite3.cmake
Remove a section from FindSqlite3.cmake to search in custom location.
There a couple of reasons for that. The first one is that cmake already
provides means to specify additional search paths. The second is an
issue of priorities -- they were wrong, so cmake ended up selecting host
sqlite3 when I tried to cross-compile.